Fix counter overflows when comparing 10000+ files
- Change Prefs::_numberOfVideos from int to int64_t to handle large video counts
- Update all calculations using _numberOfVideos * (_numberOfVideos-1)/2 to use int64_t to prevent overflow
- Update comparisonsSoFar() function to use int64_t for all counter variables
- Change foundMatches and numScanned in reportMatchingVideos() to int64_t
- Update UI components to handle int64_t values properly
- Add test for large number calculations to prevent regression

Fixes #165

+++ QtProject/app/comparison.h
@@ -106,7 +106,7 @@ private:
     // --- //
 
 public slots:
-    int reportMatchingVideos(); // returns number of matching videos found
+    int64_t reportMatchingVideos(); // returns number of matching videos found
 
 private slots:
     void dragEnterEvent(QDragEnterEvent *event); // drag and drop for locked folders list
@@ -124,7 +124,7 @@ private slots:
     QString readableBitRate(const double &kbps) const;
     void highlightBetterProperties() const;
     void updateUI();
-    int comparisonsSoFar() const;
+    int64_t comparisonsSoFar() const;
     void onProgressSliderReleased();
 
     void on_selectPhash_clicked ( const bool &checked) { if(checked) this->_prefs.comparisonMode(Prefs::_PHASH);
